servo: Merge #531 - Dynamic Resolution (from eschweic:master); r=metajack
There is now a SharedChan from the compositor to the renderer. Using this, tiles are re-rendered the tiles at a higher resolution after zooming (this is currently bound to a MouseClickUp event). Also, extra texture layers are deleted, which fixes a bug where old content was being displayed when a new paint message was sent.

Depends on rust-layers: 1e1e359da32c9d0ddb4f93b505a658d669008426

Source-Repo: https://github.com/servo/servo
Source-Revision: 8f32094d73f29f50387362fc6cee730344fec379

UltraBlame original commit: 68734339c5be27500dec8c2d0c86e354f6f183c7

+++ servo/src/components/msg/compositor.rs
@@ -11,8 +11,9 @@ pub struct LayerBuffer {
     draw_target: DrawTarget,
 
     
-    rect: Rect<uint>,
+    rect: Rect<f32>,
 
+    
     screen_pos: Rect<uint>,
